+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/combobox/DefaultFilterComboBoxModel.java
@@ -20,8 +20,8 @@ package is.codion.swing.common.model.component.combobox;
 
 import is.codion.common.Text;
 import is.codion.common.event.Event;
-import is.codion.common.model.FilterModel.Items.FilteredItems;
-import is.codion.common.model.FilterModel.Items.VisibleItems;
+import is.codion.common.model.FilterModel.Items.Filtered;
+import is.codion.common.model.FilterModel.Items.Visible;
 import is.codion.common.observer.Mutable;
 import is.codion.common.observer.Observer;
 import is.codion.common.state.State;
@@ -337,8 +337,8 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 						.listener(this::filter)
 						.build();
 
-		private final DefaultVisibleItems visible = new DefaultVisibleItems();
-		private final DefaultFilteredItems filtered = new DefaultFilteredItems();
+		private final VisibleItems visible = new VisibleItems();
+		private final FilteredItems filtered = new FilteredItems();
 
 		private final Event<Collection<T>> event = Event.event();
 
@@ -378,12 +378,12 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 		}
 
 		@Override
-		public VisibleItems<T> visible() {
+		public Visible<T> visible() {
 			return visible;
 		}
 
 		@Override
-		public FilteredItems<T> filtered() {
+		public Filtered<T> filtered() {
 			return filtered;
 		}
 
@@ -426,7 +426,7 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 		}
 	}
 
-	private final class DefaultVisibleItems implements VisibleItems<T> {
+	private final class VisibleItems implements Visible<T> {
 
 		private final List<T> items = new ArrayList<>();
 		private final Event<List<T>> event = Event.event();
@@ -484,7 +484,7 @@ class DefaultFilterComboBoxModel<T> implements FilterComboBoxModel<T> {
 		}
 	}
 
-	private final class DefaultFilteredItems implements FilteredItems<T> {
+	private final class FilteredItems implements Filtered<T> {
 
 		private final List<T> items = new ArrayList<>();
 		private final Event<Collection<T>> event = Event.event();